% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{checkup_GeneTonic}
\alias{checkup_GeneTonic}
\title{Checking the input objects for GeneTonic}
\usage{
checkup_GeneTonic(dds, res_de, res_enrich, annotation_obj)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object, normally obtained after running your data
through the \code{DESeq2} framework.}

\item{res_de}{A \code{DESeqResults} object. As for the \code{dds} parameter, this is
also commonly used in the \code{DESeq2} framework.}

\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{annotation_obj}{A \code{data.frame} object, containing two columns, \code{gene_id}
with a set of unambiguous identifiers (e.g. ENSEMBL ids) and \code{gene_name},
containing e.g. HGNC-based gene symbols.}
}
\value{
Invisible NULL
}
\description{
Checking the input objects for GeneTonic, whether these are all set for running
the app
}
\details{
Some suggestions on the requirements for each parameter are returned in the
error messages.
}
\examples{
#
}
