% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_distances.R
\name{create_jaccard_matrix}
\alias{create_jaccard_matrix}
\title{Compute the overlap matrix for enrichment results}
\usage{
create_jaccard_matrix(
  res_enrich,
  n_gs = nrow(res_enrich),
  gs_ids = NULL,
  return_sym = FALSE
)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to see the
formatting requirements.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be included (from the top ranked ones). Defaults to the number of rows of
\code{res_enrich}}

\item{gs_ids}{Character vector, containing a subset of \code{gs_id} as they are
available in \code{res_enrich}. Lists the gene sets to be included, additionally to
the ones specified via \code{n_gs}. Defaults to NULL.}

\item{return_sym}{Logical, whether to return the symmetrical matrix or just the
upper triangular - as needed by \code{\link[=enrichment_map]{enrichment_map()}}, for example.}
}
\value{
A matrix with the kappa scores between gene sets
}
\description{
Compute the overlap matrix for enrichment results, based on the Jaccard Index
between each pair of sets
}
\examples{
# TODO
}
\seealso{
\code{\link[=gs_mds]{gs_mds()}}, \code{\link[=enrichment_map]{enrichment_map()}}
}
