% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_matrix.R
\name{create_kappa_matrix}
\alias{create_kappa_matrix}
\title{Compute the kappa matrix for enrichment results}
\usage{
create_kappa_matrix(res_enrich)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to see the
formatting requirements.}
}
\value{
A matrix with the kappa scores between gene sets
}
\description{
Compute the kappa matrix for enrichment results, based on the overlap or the
Jaccard Index between each pair of sets
}
\examples{
# TODO
}
\seealso{
\code{\link[=gs_mds]{gs_mds()}}
}
