% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_hour.R
\name{.check_pandoc}
\alias{.check_pandoc}
\title{Check whether \code{pandoc} and \code{pandoc-citeproc} are available}
\usage{
.check_pandoc(ignore_pandoc)
}
\arguments{
\item{ignore_pandoc}{Logical. If TRUE, just give a warning if one of pandoc or
pandoc-citeproc is not available. If FALSE, an error is thrown.}
}
\value{
No value is returned. If \code{pandoc} or \code{pandoc-citeproc} are missing,
either warning or error messages are triggered.
}
\description{
Check whether \code{pandoc} and \code{pandoc-citeproc} are available
}
\details{
Credits to the original implementation proposed by Charlotte Soneson,
upon which this function is \strong{heavily} inspired.
}
