% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich2graph.R
\name{enrich2graph}
\alias{enrich2graph}
\title{Title}
\usage{
enrich2graph(res_enrich, res_de, annotation_obj = NULL, n_gs = 15,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID", prettify = TRUE,
  geneset_graph_color = "gold", genes_graph_colpal)
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{annotation_obj}{TODO}

\item{n_gs}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{prettify}{TODO}

\item{geneset_graph_color}{TODO}

\item{genes_graph_colpal}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
#TODO
}
