% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-data.R
\docType{data}
\name{enrichr_output_macrophage}
\alias{enrichr_output_macrophage}
\title{A sample output from Enrichr}
\description{
A sample output object as created from a call to Enrichr, with the interface
provided by \code{enrichR} - using the \code{enrichr()} function
}
\details{
This object has been created on the data from the \code{macrophage} package
by analyzing downstream the differentially expressed genes
when comparing IFNg treated samples vs naive samples, accounting
for the different cell lines included.

Details on how this object has been created are included in the \code{create_gt_data.R}
script, included in the \code{scripts} folder of the \code{GeneTonic} package.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
\seealso{
Other pathway-analysis-results:
\code{\link{gostres_macrophage}},
\code{\link{topgoDE_macrophage_IFNg_vs_naive}}
}
\concept{pathway-analysis-results}
