% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-pkg.R
\docType{package}
\name{GeneTonic-pkg}
\alias{GeneTonic-package}
\alias{GeneTonic-pkg}
\title{GeneTonic}
\description{
\code{GeneTonic} is a Bioconductor package that provides an interactive Shiny-based
graphical user interface for streamlining the interpretation of RNA-seq data
}
\details{
\code{GeneTonic} simplifies and optimizes the integration of all components of
Differential Expression analysis, with functional enrichment analysis and the
original expression quantifications.
It does so in a way that makes it easier to generate insightful observations
and hypothesis - combining the benefits of interactivity and reproducibility,
e.g. by capturing the features and gene sets of interest highlighted during
the live session, and creating an HTML report as an artifact where text,
code, and output coexist.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/GeneTonic}
  \item Report bugs at \url{https://github.com/federicomarini/GeneTonic/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Annekathrin Nedwed \email{anneludt@uni-mainz.de} (\href{https://orcid.org/0000-0002-2475-4945}{ORCID})
}

Other contributors:
\itemize{
  \item Edoardo Filippi \email{edoardo.filippi@uni-mainz.de} [contributor]
}

}
\keyword{internal}
