% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic.R
\name{GeneTonic}
\alias{GeneTonic}
\title{GeneTonic}
\usage{
GeneTonic(dds, res, func_enrich)
}
\arguments{
\item{dds}{dds}

\item{res}{result}

\item{func_enrich}{functional enrichment}
}
\value{
A Shiny app object is returned, for interactive data exploration
}
\description{
GeneTonic, main function for the Shiny app
}
