% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_table.R
\name{get_aggrscores}
\alias{get_aggrscores}
\title{Title TODO
TODO}
\usage{
get_aggrscores(res_enrich, res_de, n_gs = 50, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  annotation_obj, aggrfun = mean)
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{n_gs}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{annotation_obj}{TODO}

\item{aggrfun}{TODO}
}
\value{
TODO
}
\description{
Title TODO
TODO
}
\examples{
# TODO

# TODOTODO: add some standardized column names, say "count"
}
