% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_mds.R
\name{go_mds}
\alias{go_mds}
\title{Title TODO}
\usage{
go_mds(res_enrich, res_de, annotation_obj, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  genes_separator = ",", similarity_measure = "kappa_matrix",
  mds_k = 2, mds_labels = 10, mds_colorby = "z_score")
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{annotation_obj}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{genes_separator}{TODO}

\item{similarity_measure}{TODO}

\item{mds_k}{TODO}

\item{mds_labels}{TODO}

\item{mds_colorby}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
