% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_alluvial.R
\name{gs_alluvial}
\alias{gs_alluvial}
\title{Title TODO}
\usage{
gs_alluvial(res_enrich, res_de, annotation_obj, n_gs = 5,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID")
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{annotation_obj}{TODO}

\item{n_gs}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
