% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_ggheatmap}
\alias{gs_ggheatmap}
\title{Title}
\usage{
gs_ggheatmap(mat, clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean", cluster_rows = TRUE,
  cluster_cols = TRUE)
}
\arguments{
\item{mat}{TODO}

\item{clustering_distance_rows}{TODO}

\item{clustering_distance_cols}{TODO}

\item{cluster_rows}{TODO}

\item{cluster_cols}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
