% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_heatmap}
\alias{gs_heatmap}
\title{Title}
\usage{
gs_heatmap(se, res_de, res_enrich, geneset_id, genelist,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID", annotation_obj = NULL, FDR = 0.05,
  de_only = FALSE, cluster_rows = TRUE, cluster_cols = FALSE,
  center_mean = TRUE, scale_row = FALSE)
}
\arguments{
\item{scale_row}{}
}
\description{
Title
}
