% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_radar.R
\name{gs_radar}
\alias{gs_radar}
\title{Title TODO}
\usage{
gs_radar(res_enrich1, res_enrich2, n_gs = 20,
  p_value_column = "p.value_elim")
}
\arguments{
\item{res_enrich1}{TODO}

\item{res_enrich2}{TODO}

\item{n_gs}{TODO}

\item{p_value_column}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
