% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_scores}
\alias{gs_scores}
\title{Title}
\usage{
gs_scores(se, res_de, res_enrich, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  annotation_obj = NULL)
}
\arguments{
\item{se}{TODO}

\item{res_de}{TODO}

\item{res_enrich}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{annotation_obj}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
