% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_summaries.R
\name{gs_summary_heat}
\alias{gs_summary_heat}
\title{Title}
\usage{
gs_summary_heat(res_enrich, res_de, n_gs = 80, annotation_obj,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID")
}
\arguments{
\item{genesetid_colname}{}
}
\description{
Title
}
