% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_summaries.R
\name{gs_summary_overview_pair}
\alias{gs_summary_overview_pair}
\title{Title  TODO}
\usage{
gs_summary_overview_pair(res_enrich, res_enrich2, n_gs = 20,
  p_value_column = "p.value_elim", color_by = "z_score",
  alpha_set2 = 0.4)
}
\arguments{
\item{res_enrich}{TODO}

\item{res_enrich2}{TODO}

\item{n_gs}{TODO}

\item{p_value_column}{TODO}

\item{color_by}{TODO}

\item{alpha_set2}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
