% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_upset.R
\name{gs_upset}
\alias{gs_upset}
\title{Upset plot for genesets}
\usage{
gs_upset(
  res_enrich,
  res_de,
  annotation_obj = NULL,
  n_gs = 10,
  gtl = NULL,
  gs_ids = NULL,
  add_de_direction = FALSE,
  add_de_gsgenes = FALSE,
  col_upDE = "#E41A1C",
  col_downDE = "#377EB8",
  upset_geom = geom_point(size = 2),
  return_upsetgsg = FALSE
)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be included}

\item{gtl}{A \code{GeneTonic}-list object, containing in its slots the arguments
specified above: \code{dds}, \code{res_de}, \code{res_enrich}, and \code{annotation_obj} - the names
of the list \emph{must} be specified following the content they are expecting}

\item{gs_ids}{Character vector, containing a subset of \code{gs_id} as they are
available in \code{res_enrich}. Lists the gene sets to be included in addition to
the top ones (via \code{n_gs})}

\item{add_de_direction}{Logical, whether to add an annotation with info on the
DE direction of single genes}

\item{add_de_gsgenes}{Logical, if set to TRUE adds an annotation with detail
on the single components of each defined subset}

\item{col_upDE}{Character, specifying the color value to be used to mark
upregulated genes}

\item{col_downDE}{Character, specifying the color value to be used to mark
downregulated genes}

\item{upset_geom}{A geom specification to be used in the upset chart. Defaults
sensibly to \code{geom_point(size = 2)}}

\item{return_upsetgsg}{Logical, controlling the returned value. If set to TRUE,
this function will not generate the plot but only create the corresponding
data.frame, in case the user wants to proceed with a custom call to create an
upset plot.}
}
\value{
A \code{ggplot} object (if plotting), or alternatively a data.frame
}
\description{
Create an upset plot for genesets
}
\examples{
# TODO
}
