% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_volcano.R
\name{gs_volcano}
\alias{gs_volcano}
\title{Title TODO}
\usage{
gs_volcano(res_enrich, labels_to_use = "Term",
  pvals_to_use = "p.value_elim", p_threshold = 0.05,
  max_nr_labels = 10, scale_circles = 1)
}
\arguments{
\item{res_enrich}{TODO}

\item{labels_to_use}{TODO}

\item{pvals_to_use}{TODO}

\item{p_threshold}{TODO}

\item{max_nr_labels}{TODO}

\item{scale_circles}{TODO}
}
\value{
TODO
}
\description{
Title TODO
}
\examples{
#  TODO
}
