% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-extras.R
\name{overlap_jaccard_index}
\alias{overlap_jaccard_index}
\title{Calculate Jaccard Index between two sets}
\usage{
overlap_jaccard_index(x, y)
}
\arguments{
\item{x}{Character vector, corresponding to set 1}

\item{y}{Character vector, corresponding to set 2}
}
\value{
A numeric value between 0 and 1
}
\description{
Calculate similarity coefficient with the Jaccard Index
}
\examples{
a <- seq(1, 21, 2)
b <- seq(1, 11, 2)
overlap_jaccard_index(a, b)
}
