% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_shaker.R
\name{shake_fgseaResult}
\alias{shake_fgseaResult}
\title{Convert the output of fgsea}
\usage{
shake_fgseaResult(fgsea_output)
}
\arguments{
\item{fgsea_output}{A data.frame with the output of \code{fgsea()} in \code{fgsea}.}
}
\value{
A \code{data.frame} compatible for use in \code{\link[=GeneTonic]{GeneTonic()}} as \code{res_enrich}
}
\description{
Convert the output of fgsea for straightforward use in \code{\link[=GeneTonic]{GeneTonic()}}
}
\examples{
data(fgseaRes, package = "GeneTonic")
res_from_fgsea <- shake_fgseaResult(fgseaRes)
}
\seealso{
Other shakers:
\code{\link[=shake_davidResult]{shake_davidResult()}},
\code{\link[=shake_enrichResult]{shake_enrichResult()}},
\code{\link[=shake_enrichrResult]{shake_enrichrResult()}},
\code{\link[=shake_gprofilerResult]{shake_gprofilerResult()}},
\code{\link[=shake_gsenrichResult]{shake_gsenrichResult()}},
\code{\link[=shake_topGOtableResult]{shake_topGOtableResult()}}
}
\concept{shakers}
