% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cluster.R
\name{cluster_markov}
\alias{cluster_markov}
\title{Markov Clustering (MCL) for community detection}
\usage{
cluster_markov(
  g,
  add_self_loops = TRUE,
  loop_value = 1,
  mcl_expansion = 2,
  mcl_inflation = 2,
  allow_singletons = TRUE,
  max_iter = 100,
  return_node_names = TRUE,
  return_esm = FALSE
)
}
\arguments{
\item{g}{The input graph object}

\item{add_self_loops}{TODO}

\item{loop_value}{TODO}

\item{mcl_expansion}{TODO}

\item{mcl_inflation}{TODO}

\item{allow_singletons}{TODO}

\item{max_iter}{TODO}

\item{return_node_names}{TODO}

\item{return_esm}{TODO}
}
\value{
This function returns a \code{communities} object, containing the numbers of
the assigned membership. Please see the \code{\link[igraph:communities]{igraph::communities()}} manual page for details
}
\description{
This function implements the Markov Clustering (MCL) algorithm for finding community
structure, in an analogous way to other existing algorithms in \code{igraph}.
}
\details{
This implementation has been driven by the nice explanations provided in
\itemize{
\item https://sites.cs.ucsb.edu/~xyan/classes/CS595D-2009winter/MCL_Presentation2.pdf
\item https://medium.com/analytics-vidhya/demystifying-markov-clustering-aeb6cdabbfc7
\item https://github.com/GuyAllard/markov_clustering (python implementation)
}

More info on the MCL: https://micans.org/mcl/index.html, and
https://micans.org/mcl/sec_description1.html
}
\examples{
# TODO
}
\references{
van Dongen, S.M., Graph clustering by flow simulation (2000) PhD thesis,
Utrecht University Repository - https://dspace.library.uu.nl/handle/1874/848

Enright AJ, van Dongen SM, Ouzounis CA, An efficient algorithm for
large-scale detection of protein families (2002) Nucleic Acids Research, Volume 30,
Issue 7, 1 April 2002, Pages 1575–1584, https://doi.org/10.1093/nar/30.7.1575
}
