% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_matrix.R
\name{create_kappa_matrix}
\alias{create_kappa_matrix}
\title{Compute the kappa matrix for enrichment results}
\usage{
create_kappa_matrix(res_enrich, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  genes_separator = ",")
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to see the
formatting requirements.}

\item{genes_colname}{Character, specifying which column of the \code{res_enrich}
object contains the genes assigned to each gene set, detected as differentially
expressed. Defaults to \code{genes}.}

\item{genesetname_colname}{Character, specifies which column of the \code{res_enrich}
object contains a description of the gene set. Defaults to \code{Term}.}

\item{genesetid_colname}{Character, specifies which column of the \code{res_enrich}
object contains a unique identifier of the gene set. Defaults to \code{GO.ID}.}

\item{genes_separator}{TODO}
}
\value{
A matrix with the kappa scores between gene sets
}
\description{
Compute the kappa matrix for enrichment results, based on the overlap or the
Jaccard Index between each pair of sets
}
\examples{
# TODO
}
\seealso{
\code{\link[=gs_mds]{gs_mds()}}
}
