% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_matrix.R
\name{create_kappa_matrix}
\alias{create_kappa_matrix}
\title{Title TODO}
\usage{
create_kappa_matrix(res_enrich, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  genes_separator = ",")
}
\arguments{
\item{res_enrich}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{genes_separator}{TODO}
}
\value{
TODO
}
\description{
Title TODO
}
\examples{
# TODO
}
