% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-extras.R
\name{deseqresult2df}
\alias{deseqresult2df}
\title{Title TODO}
\usage{
deseqresult2df(deseqresult, FDR = NULL)
}
\arguments{
\item{deseqresult}{TODO}

\item{FDR}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
