% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_table.R
\name{enhance_table}
\alias{enhance_table}
\title{Title}
\usage{
enhance_table(res_enrich, res_de, annotation_obj, n_gs = 50,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID", chars_limit = 60)
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{annotation_obj}{TODO}

\item{n_gs}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{chars_limit}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
#TODO
}
