% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_table.R
\name{enhance_table}
\alias{enhance_table}
\title{Visually enhances a functional enrichment result table}
\usage{
enhance_table(res_enrich, res_de, annotation_obj, n_gs = 50,
  gs_ids = NULL, chars_limit = 60)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation.
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be displayed.}

\item{gs_ids}{Character vector, containing a subset of \code{gs_id} as they are
available in \code{res_enrich}. Lists the gene sets to be displayed.}

\item{chars_limit}{Integer, number of characters to be displayed for each
geneset name.}
}
\value{
A \code{ggplot} object
}
\description{
Creates a visual summary for the results of a functional enrichment analysis,
by displaying also the components of each gene set and their expression change
in the contrast of interest
}
\examples{
#TODO
}
