% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich2graph.R
\name{enrich2graph}
\alias{enrich2graph}
\title{Title}
\usage{
enrich2graph(res_enrich, res_de, n_nodes = 15, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID",
  prettify = TRUE, geneset_graph_color = "gold", genes_graph_colpal,
  annotation_obj = NULL)
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{n_nodes}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{prettify}{TODO}

\item{geneset_graph_color}{TODO}

\item{genes_graph_colpal}{TODO}

\item{annotation_obj}{TODO}
}
\description{
Title
}
