% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_map.R
\name{enrichment_map}
\alias{enrichment_map}
\title{Title}
\usage{
enrichment_map(res_enrich, res_de, annotation_obj, n_gs = 50,
  overlap_threshold = 0.1, scale_edges_width = 200,
  color_by = "p.value_elim", size_by, genes_colname = "genes",
  genesetname_colname = "Term", genesetid_colname = "GO.ID")
}
\arguments{
\item{res_enrich}{TODO}

\item{res_de}{TODO}

\item{annotation_obj}{TODO}

\item{n_gs}{TODO}

\item{overlap_threshold}{TODO}

\item{scale_edges_width}{TODO}

\item{color_by}{TODO}

\item{size_by}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO

TODOTODO: similarity measures, say, jaccard, or simple overlap}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
