% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{gene_plot}
\alias{gene_plot}
\title{Title TODO}
\usage{
gene_plot(dds, gene, assay = "counts", intgroup = "condition",
  annotation_obj = NULL, transform = TRUE, labels_repel = TRUE,
  plot_type = "auto")
}
\arguments{
\item{dds}{TODO}

\item{gene}{TODO}

\item{assay}{TODO}

\item{intgroup}{TODO}

\item{annotation_obj}{TODO}

\item{transform}{TODO}

\item{labels_repel}{TODO}

\item{plot_type}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
