% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-pkg.R
\docType{package}
\name{GeneTonic-pkg}
\alias{GeneTonic-pkg}
\alias{GeneTonic-pkg-package}
\title{GeneTonic}
\description{
\code{GeneTonic} is a Bioconductor package that provides an interactive Shiny-based
graphical user interface for...
}
\author{
Federico Marini \email{marinif@uni-mainz.de}

Simon Tiado Stahl \email{simon.t.stahl@gmail.com}
}
