% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-pkg.R
\docType{package}
\name{GeneTonic-pkg}
\alias{GeneTonic-pkg}
\title{GeneTonic}
\description{
\code{GeneTonic} is a Bioconductor package that provides an interactive Shiny-based
graphical user interface for streamlining the interpretation of RNA-seq data

\code{GeneTonic} simplifies and optimizes the integration of all components of
Diffeential Expression analysis, with functional enrichment analyis and the
original expression quantifications.
It does so in a way that makes it easier to generate insightful observations
and hypothesis - combining the benefits of interactivity and reproducibility,
e.g. by capturing the features and gene sets of interest highlighted during
the live session, and creating an HTML report as an artifact where text,
code, and output coexist.
}
\author{
Federico Marini \email{marinif@uni-mainz.de}
}
