% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic.R
\name{GeneTonic}
\alias{GeneTonic}
\title{GeneTonic}
\usage{
GeneTonic(dds, res_de, res_enrich)
}
\arguments{
\item{dds}{dds}

\item{res_de}{res_de}

\item{res_enrich}{res_enrich}
}
\value{
A Shiny app object is returned, for interactive data exploration
}
\description{
GeneTonic, main function for the Shiny app
}
\examples{
library(GeneTonic)
# whatever comes next
}
\author{
Federico Marini
}
