% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-extras.R
\name{go_2_html}
\alias{go_2_html}
\title{Information on a GeneOntology identifier}
\usage{
go_2_html(go_id, res_enrich = NULL)
}
\arguments{
\item{go_id}{Character, specifying the GeneOntology identifier for which
to retrieve information}

\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. If not provided, the experiment-related information is not
shown, and only some generic info on the identifier is displayed.
See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}
}
\value{
HTML content related to a GeneOntology identifier, to be displayed in
web applications (or inserted in Rmd documents)
}
\description{
Assembles information, in HTML format, regarding a Gene Ontology identifier
}
\details{
Also creates a link to the AmiGO database
}
\examples{
go_2_html("GO:0002250")
go_2_html("GO:0043368")
}
