% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_alluvial.R
\name{gs_alluvial}
\alias{gs_alluvial}
\title{Alluvial (sankey) plot for a set of genesets and the associated genes}
\usage{
gs_alluvial(res_enrich, res_de, annotation_obj, n_gs = 5, gs_ids = NULL)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be displayed}

\item{gs_ids}{Character vector, containing a subset of \code{gs_id} as they are
available in \code{res_enrich}. Lists the gene sets to be displayed.}
}
\value{
A \code{plotly} object
}
\description{
Generate an interactive alluvial plot linking genesets to their associated
genes
}
\examples{
# TODO
}
