% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_heatmap}
\alias{gs_heatmap}
\title{Title}
\usage{
gs_heatmap(se, res_de, res_enrich, annotation_obj = NULL, geneset_id,
  genelist, genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID", FDR = 0.05, de_only = FALSE,
  cluster_rows = TRUE, cluster_cols = FALSE, center_mean = TRUE,
  scale_row = FALSE)
}
\arguments{
\item{se}{TODO}

\item{res_de}{TODO}

\item{res_enrich}{TODO}

\item{annotation_obj}{TODO}

\item{geneset_id}{TODO}

\item{genelist}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}

\item{FDR}{TODO}

\item{de_only}{TODO}

\item{cluster_rows}{TODO}

\item{cluster_cols}{TODO}

\item{center_mean}{TODO}

\item{scale_row}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
