% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_summaries.R
\name{gs_horizon}
\alias{gs_horizon}
\title{Title TODO}
\usage{
gs_horizon(res_enrich, n_gs = 20, p_value_column = "p.value_elim",
  color_by = "z_score")
}
\arguments{
\item{res_enrich}{TODO}

\item{n_gs}{TODO}

\item{p_value_column}{TODO}

\item{color_by}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
