% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_summaries.R
\name{gs_horizon}
\alias{gs_horizon}
\title{Plots a summary of enrichment results}
\usage{
gs_horizon(
  res_enrich,
  n_gs = 20,
  p_value_column = "gs_pvalue",
  color_by = "z_score"
)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be displayed}

\item{p_value_column}{TODO}

\item{color_by}{Character, specifying the column of \code{res_enrich} to be used
for coloring the plotted gene sets. Defaults sensibly to \code{z_score}.}
}
\value{
A \code{ggplot} object
}
\description{
Plots a summary of enrichment results - horizon plot to compare one or more
sets of results
}
\examples{
# TODO
}
\seealso{
\code{\link[=gs_summary_overview]{gs_summary_overview()}}, \code{\link[=gs_summary_overview_pair]{gs_summary_overview_pair()}}
}
