% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_mds.R
\name{gs_mds}
\alias{gs_mds}
\title{Multi Dimensional Scaling plot for gene sets}
\usage{
gs_mds(
  res_enrich,
  res_de,
  annotation_obj,
  similarity_measure = "kappa_matrix",
  mds_k = 2,
  mds_labels = 0,
  mds_colorby = "z_score",
  gs_labels = NULL
)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{similarity_measure}{Character, currently defaults to \code{kappa_matrix}, to
specify how to compute the similarity measure between gene sets}

\item{mds_k}{Integer value, number of dimensions to compute in the multi
dimensional scaling procedure}

\item{mds_labels}{Integer, defines the number of labels to be plotted on top
of the scatter plot for the provided gene sets.}

\item{mds_colorby}{Character specifying the column of \code{res_enrich} to be used
for coloring the plotted gene sets. Defaults sensibly to \code{z_score}.}

\item{gs_labels}{Character vector, containing a subset of \code{gs_id} as they are
available in \code{res_enrich}. Lists the gene sets to be labelled.}
}
\value{
A \code{ggplot} object
}
\description{
Multi Dimensional Scaling plot for gene sets, extracted from a \code{res_enrich}
object
}
\examples{
# TODO
}
\seealso{
\code{\link[=create_kappa_matrix]{create_kappa_matrix()}} is used to calculate the similarity between
gene sets
}
