% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_scores}
\alias{gs_scores}
\title{Title}
\usage{
gs_scores(se, res_de, res_enrich, annotation_obj = NULL,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID")
}
\arguments{
\item{se}{TODO}

\item{res_de}{TODO}

\item{res_enrich}{TODO}

\item{annotation_obj}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genesetid_colname}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
