% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heatmap.R
\name{gs_scoresheat}
\alias{gs_scoresheat}
\title{Plots a matrix of geneset scores}
\usage{
gs_scoresheat(
  mat,
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  cluster_rows = TRUE,
  cluster_cols = TRUE
)
}
\arguments{
\item{mat}{A matrix, e.g. returned by the \code{\link[=gs_scores]{gs_scores()}} function}

\item{clustering_distance_rows}{Character, a distance measure used in
clustering rows}

\item{clustering_distance_cols}{Character, a distance measure used in
clustering columns}

\item{cluster_rows}{Logical, determining if rows should be clustered}

\item{cluster_cols}{Logical, determining if columns should be clustered}
}
\value{
A \code{ggplot} object
}
\description{
Plots a matrix of geneset Z scores, across all samples
}
\examples{
# TODO
}
\seealso{
\code{\link[=gs_scores]{gs_scores()}} computes the scores plotted by this function
}
