% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_simplify.R
\name{gs_simplify}
\alias{gs_simplify}
\title{Title TODO}
\usage{
gs_simplify(res_enrich, gs_overlap = 0.75, genes_colname = "genes",
  genesetname_colname = "Term", genes_separator = ",")
}
\arguments{
\item{res_enrich}{TODO}

\item{gs_overlap}{TODO}

\item{genes_colname}{TODO}

\item{genesetname_colname}{TODO}

\item{genes_separator}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
