% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_summaries.R
\name{gs_summary_heat}
\alias{gs_summary_heat}
\title{Plots a heatmap for genes and genesets}
\usage{
gs_summary_heat(res_enrich, res_de, annotation_obj, n_gs = 80)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to check the
formatting requirements (a minimal set of columns should be present).}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be displayed}
}
\value{
A \code{ggplot} object
}
\description{
Plots a heatmap for genes and genesets, useful to spot out intersections across
genesets and an overview of them
}
\examples{
# TODO
}
