% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_summaries.R
\name{gs_summary_heat}
\alias{gs_summary_heat}
\title{Plots a heatmap for genes and genesets}
\usage{
gs_summary_heat(res_enrich, res_de, annotation_obj, n_gs = 80,
  genes_colname = "genes", genesetname_colname = "Term",
  genesetid_colname = "GO.ID")
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to see the
formatting requirements.}

\item{res_de}{A \code{DESeqResults} object.}

\item{annotation_obj}{A \code{data.frame} object with the feature annotation
information, with at least two columns, \code{gene_id} and \code{gene_name}.}

\item{n_gs}{Integer value, corresponding to the maximal number of gene sets to
be displayed}

\item{genes_colname}{Character, specifying which column of the \code{res_enrich}
object contains the genes assigned to each gene set, detected as differentially
expressed. Defaults to \code{genes}.}

\item{genesetname_colname}{Character, specifies which column of the \code{res_enrich}
object contains a description of the gene set. Defaults to \code{Term}.}

\item{genesetid_colname}{Character, specifies which column of the \code{res_enrich}
object contains a unique identifier of the gene set. Defaults to \code{GO.ID}.}
}
\value{
A \code{ggplot} object
}
\description{
Plots a heatmap for genes and genesets, useful to spot out intersections across
genesets and an overview of them
}
\examples{
# TODO
}
