% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_volcano.R
\name{gs_volcano}
\alias{gs_volcano}
\title{Volcano plot for gene sets}
\usage{
gs_volcano(res_enrich, p_threshold = 0.05, max_nr_labels = 10,
  scale_circles = 1)
}
\arguments{
\item{res_enrich}{A \code{data.frame} object, storing the result of the functional
enrichment analysis. See more in the main function, \code{\link[=GeneTonic]{GeneTonic()}}, to see the
formatting requirements. This object needs to be processed first by a function
such as \code{\link[=get_aggrscores]{get_aggrscores()}} to compute the term-wise \code{z_score} or \code{aggr_score},
which will be used for plotting}

\item{p_threshold}{Numeric, defines the threshold to be used for filtering the
gene sets to display. Defaults to 0.05}

\item{max_nr_labels}{Integer, maximum number of labels for the gene sets to be
plotted as labels on the volcano scatter plot.}

\item{scale_circles}{TODO}
}
\value{
A \code{ggplot} object
}
\description{
Volcano plot for gene sets, to summarize visually the functional enrichment
results
}
\examples{
#  TODO
}
\seealso{
\code{\link[=gs_simplify]{gs_simplify()}} can be applied in advance to \code{res_enrich} to reduce
the redundancy of the displayed gene sets
}
