% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-extras.R
\name{map2color}
\alias{map2color}
\title{Title}
\usage{
map2color(x, pal, limits = NULL)
}
\arguments{
\item{x}{TODO}

\item{pal}{TODO}

\item{limits}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
