% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-extras.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Calculate overlap coefficient}
\usage{
overlap_coefficient(x, y)
}
\arguments{
\item{x}{Character vector, corresponding to set 1}

\item{y}{Character vector, set 2}
}
\value{
A numeric value between 0 and 1
}
\description{
Calculate similarity coefficient between two sets, based on the overlap
}
\examples{
a <- seq(1, 21, 2)
b <- seq(1, 11, 2)
overlap_coefficient(a,b)
}
\seealso{
https://en.wikipedia.org/wiki/Overlap_coefficient
}
