% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_map_static.R
\name{plot_emap_static}
\alias{plot_emap_static}
\title{Creates the plot of the enrichment map graph}
\usage{
plot_emap_static(
  emg,
  scale_edges_width = 5,
  scale_nodes_size = 10,
  cluster_fun = "cluster_markov"
)
}
\arguments{
\item{emg}{An \code{igraph} object, ideally generated via \code{enrichment_map()}}

\item{scale_edges_width}{A numeric value, to define the scaling factor for the
edges between nodes. Defaults to 200 (works well chained to \code{visNetwork}
functions).}

\item{scale_nodes_size}{A numeric value, to define the scaling factor for the
within GeneTonic}

\item{cluster_fun}{Character string, containing the name of the function
to be used for clustering the graph of the geneset similarities. Defaults to
"cluster_markov".}
}
\value{
A \code{ggraph} object with the static representation of the enrichment
map graph.
}
\description{
Generates a graph plot for the enrichment map, as created within \code{GeneTonic}
via \code{enrichment_map()}.
}
\examples{
library("macrophage")
library("DESeq2")
library("org.Hs.eg.db")
library("AnnotationDbi")

# dds object
data("gse", package = "macrophage")
dds_macrophage <- DESeqDataSet(gse, design = ~ line + condition)
rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)
dds_macrophage <- estimateSizeFactors(dds_macrophage)

# annotation object
anno_df <- data.frame(
  gene_id = rownames(dds_macrophage),
  gene_name = mapIds(org.Hs.eg.db,
    keys = rownames(dds_macrophage),
    column = "SYMBOL",
    keytype = "ENSEMBL"
  ),
  stringsAsFactors = FALSE,
  row.names = rownames(dds_macrophage)
)

# res object
data(res_de_macrophage, package = "GeneTonic")
res_de <- res_macrophage_IFNg_vs_naive

# res_enrich object
data(res_enrich_macrophage, package = "GeneTonic")
res_enrich <- shake_topGOtableResult(topgoDE_macrophage_IFNg_vs_naive)
res_enrich <- get_aggrscores(res_enrich, res_de, anno_df)

em <- enrichment_map(res_enrich,
  res_de,
  anno_df,
  n_gs = 50
)

em2 <- enrichment_map(res_enrich,
  res_de,
  anno_df, color_by = "z_score",
  n_gs = 100
)

em

## TODO now works with...
gtl <- GeneTonicList(dds_macrophage, 
                     res_de = res_macrophage_IFNg_vs_naive, 
                     res_enrich = res_enrich, 
                     annotation_obj = anno_df)
### enrichment_map_static(gtl = gtl, cluster_fun = "cluster_louvain")
                     
## TODO: this should work with (for example)
plot_emap_static(em, cluster_fun = "cluster_markov")

# could be viewed interactively with
# library("visNetwork")
# library("magrittr")
# em \%>\%
#   visIgraph() \%>\%
#   visOptions(highlightNearest = list(enabled = TRUE,
#                                      degree = 1,
#                                      hover = TRUE),
#             nodesIdSelection = TRUE)
}
\seealso{
\code{\link[=enrichment_map]{enrichment_map()}} is used to generate the input igraph object.
Also, \code{\link[=GeneTonic]{GeneTonic()}} embeds an interactive visualization for the enrichment
map (based on the \code{VisNetwork} package)
}
