% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_shaker.R
\name{shake_davidResult}
\alias{shake_davidResult}
\title{Convert the output of DAVID}
\usage{
shake_davidResult(david_output_file)
}
\arguments{
\item{david_output_file}{The location of the text file output, as exported from
DAVID}
}
\value{
A \code{data.frame} compatible for use in \code{\link[=GeneTonic]{GeneTonic()}} as \code{res_enrich}
}
\description{
Convert the output of DAVID for straightforward use in \code{\link[=GeneTonic]{GeneTonic()}}
}
\examples{
david_output_file <- system.file("extdata",
  "david_output_chart_BPonly_ifng_vs_naive.txt",
  package = "GeneTonic"
)
res_enrich <- shake_davidResult(david_output_file)
}
\seealso{
Other shakers: 
\code{\link{shake_enrichResult}()},
\code{\link{shake_enrichrResult}()},
\code{\link{shake_fgseaResult}()},
\code{\link{shake_gprofilerResult}()},
\code{\link{shake_topGOtableResult}()}
}
\concept{shakers}
