% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_shaker.R
\name{shake_enrichResult}
\alias{shake_enrichResult}
\title{Convert an enrichResult object}
\usage{
shake_enrichResult(obj)
}
\arguments{
\item{obj}{An \code{enrichResult} object, obtained via \code{clusterProfiler}}
}
\value{
A \code{data.frame} compatible for use in \code{\link[=GeneTonic]{GeneTonic()}} as \code{res_enrich}
}
\description{
Convert an enrichResult object for straightforward use in \code{\link[=GeneTonic]{GeneTonic()}}
}
\examples{
# TODO
}
