% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_shaker.R
\name{shake_gprofilerResult}
\alias{shake_gprofilerResult}
\title{Convert the output of g:Profiler}
\usage{
shake_gprofilerResult(gprofiler_output_file, gprofiler_output = NULL)
}
\arguments{
\item{enrichr_output_file}{The location of the text file output, as exported from
the website of g:Profiler}

\item{enrichr_output}{A data.frame with the output created with the function
\code{gost} from \code{gprofiler2}. Usually it is the \verb{$result} element of the list returned
when calling \code{gost}}
}
\value{
A \code{data.frame} compatible for use in \code{\link[=GeneTonic]{GeneTonic()}} as \code{res_enrich}
}
\description{
Convert the output of g:Profiler for straightforward use in \code{\link[=GeneTonic]{GeneTonic()}}
}
\examples{
# degenes <- (deseqresult2df(res_macrophage_IFNg_vs_naive, FDR = 0.01)$SYMBOL)
# if called directly withín R...
# enrichr_output_macrophage <- enrichr(degenes, dbs)
# or alternatively, if downloaded from the website in tabular format
gprofiler_output_file <- system.file("extdata", 
                                   "enrichr_tblexport_IFNg_vs_naive.txt", 
                                   package = "GeneTonic")
res_from_enrichr <- shake_gprofilerResult(gprofiler_output_file = gprofiler_output_file)
}
\seealso{
Other shakers: 
\code{\link{shake_davidResult}()},
\code{\link{shake_enrichResult}()},
\code{\link{shake_enrichrResult}()},
\code{\link{shake_topGOtableResult}()}
}
\concept{shakers}
