% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonic-data.R
\docType{data}
\name{topgoDE_macrophage_IFNg_vs_naive}
\alias{topgoDE_macrophage_IFNg_vs_naive}
\title{A sample \code{res_enrich} object}
\description{
A sample \code{res_enrich} object, generated with the \code{topGOtable} function (from
the \code{pcaExplorer} package).
}
\details{
This \code{res_enrich} object on the data from the \code{macrophage} package
has been created by analyzing downstream the differentially expressed genes
when comparing IFNg treated samples vs naive samples, accounting
for the different cell lines included.

Details on how this object has been created are included in the \code{create_gt_data.R}
script, included in the \code{scripts} folder of the \code{GeneTonic} package.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
\seealso{
Other pathway-analysis-results: 
\code{\link{enrichr_output_macrophage}},
\code{\link{gostres_macrophage}}
}
\concept{pathway-analysis-results}
